examples/python-fastapi/requirements.txt
fastapi==0.115.12
python-keycloak==5.5.0
pydantic-settings==2.9.1
uvicorn==0.34.2
